pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "anchormd"
version = "0.3.0"
description = "Generate and audit CLAUDE.md files for AI coding agents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "AreteDriver"}]
keywords = ["claude", "ai", "coding-agent", "claude-code", "cursor", "windsurf"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Code Generators",
    "Typing :: Typed",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "tomli>=2.0;python_version<'3.11'",
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
server = [
    "httpx>=0.27.0",
]
drift = [
    "httpx>=0.27.0",
]
dev = [
    "pytest>=7.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
    "httpx>=0.27.0",
]

[project.scripts]
anchormd = "anchormd.cli:app"

[project.urls]
Homepage = "https://github.com/Arete-Consortium/anchormd"
Repository = "https://github.com/Arete-Consortium/anchormd"
Issues = "https://github.com/Arete-Consortium/anchormd/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
